Standardize build process
