+++ Quadruped Controller/extern/enet/jni/Application.mk
@@ -1,0 +1,3 @@
+APP_PLATFORM := android-18
+APP_OPTIM := release
+APP_MODULES := libenet
