+++ Quadruped Controller/server/src/environment.cpp
@@ -1,0 +1,9 @@
+#include <environment.h>
+
+Environment::Environment()
+{
+}
+
+Environment::~Environment()
+{
+}
